Added a new feature to the Driver start stop screen
Addd the current city feature, reverse the coordinates array when returnin to origin, let the driver select the destination.
Now the driver side sucessfylly tracks the closest city from the cordinates array  to the bus current location and updates the firestore to show the current city.It also reverces the cordinates array of the bus is retuning and the destination and origin swap.